Docker: fixed sudoers username
You have no idea what I went through to find this as this commit is
probaby a force push deleting the last 7 days of my life from Gits
history (~20 pushes).
This copy-paste error causes a error that, when googled, leads you down
a path of a recent ArchLinux image
[bug](https://bugs.archlinux.org/task/69563), that seems to be the
cause, but isn't.

I've also wrote this much of a DevOps Stackexchange question, before
realizing my mistake:

```
I've been trying to migrate my CI setup from Travis to GitHub actions as I need to run it in an Arch Linux environment. This means that I have to run it in a Docker container, and that just [refused to work][1] with Travis.

So I've looked over the GitHub [actions workflow documentation][2], and [their own guide][3] on how to run a docker container via actions. There is a [complication][4], that I think stems from this [bug][5], that I need to first install the *runc* package on the host machine - meaning I need to run scripts in both the host & the container.

  [1]: https://travis-ci.community/t/unable-to-access-file-structure-of-docker-container-when-running-in-travis/11229
  [2]: https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions
  [3]: https://docs.github.com/en/actions/creating-actions/dockerfile-support-for-github-actions
  [4]: https://github.com/niraami/dotfiles/runs/1995551118
  [5]: https://bugs.archlinux.org/task/69563
```

So yeah. Let this commit be a painful lesson...